requirements.txt
pandas>=1.0.0
matplotlib>=3.0.0
python-dotenv>=0.19.0
requests>=2.25.0
